Module cpp_to_rust_common::cpp_build_config

Types for configuring build script behavior.

Structs

CppBuildConfig
Information required to build the C++ wrapper library on every supported platform. it contains list of linked libraries, frameworks, compiler types and selected type of C++ wrapper library (shared or static). Default value of this object is set before generation of the crate using cpp_to_rust_generator::config::Config::set_cpp_build_config or cpp_build_config_mut and intended to be cross-platform.

CppBuildConfigData
Platform-specific information required to build the C++ wrapper library. This type contains one configuration item of CppBuildConfig.

CppBuildPaths
Machine-specific information required to build the C++ wrapper library. This type holds configuration properties that cannot be determined at the time of crate generation because they are always platform-dependent.

Enums

CppLibraryType
Type of a C++ library (shared or static).