Struct cpp_to_rust_common::cpp_build_config::CppBuildConfig

pub struct CppBuildConfig { /* fields omitted */ }

Information required to build the C++ wrapper library on every supported platform. it contains list of linked libraries, frameworks, compiler types and selected type of C++ wrapper library (shared or static). Default value of this object is set before generation of the crate using cpp_to_rust_generator::config::Config::set_cpp_build_config or cpp_build_config_mut and intended to be cross-platform.

In order to allow target-dependent build configuration, multiple configurations can be added to one CppBuildConfig object, each with a condition. During evaluation, each configuration item will only be used if the associated condition is currently true. All properties from all matching configuration are combined.

If this conditional evaluation is not enough, a custom build script can modify this config during build script execution using cpp_to_rust_build_tools::Config::set_cpp_build_config or cpp_build_config_mut.

Methods

impl CppBuildConfig

fn new() -> CppBuildConfig

Create an empty configuration

fn add(&mut self, condition: Condition, data: CppBuildConfigData)

Add data with condition.

fn eval(&self, target: &Target) -> Result<CppBuildConfigData>

Select all conditions that are true on target, combine all corresponding configuration items and return the result.