Module cpp_to_rust_common::log

Logger implementation

Reexports

pub use self::LoggerCategory::*;

Structs

DEFAULT_LOGGER

Logger
Logger object. One logger manages messages of all categories. It's possible to use multiple loggers independently. Use default_logger() to get global Logger instance. Note that the instance is mutex-guarded.

LoggerSettings
Specifies where the logging messages should be sent.

Enums

LoggerCategory
Logger category. Logger can be configured to save messages of each category to a separate file.

Functions

default_logger
Returns global instance of Logger.

error
Convenience method to log error messages to the default logger.

llog
Convenience method to lazy-log messages to the default logger and specified category. If messages of category are not ignored, calls the passed closure and uses its output value as a message in that category.

log
Convenience method to log messages to the default logger and specified category.

status
Convenience method to log status messages to the default logger.