Enum cpp_to_rust_common::target::Condition

pub enum Condition {
    Arch(Arch),
    OS(OS),
    Family(Family),
    Env(Env),
    PointerWidth(PointerWidth),
    Endian(Endian),
    And(Vec<Condition>),
    Or(Vec<Condition>),
    Not(Box<Condition>),
    True,
    False,
}

Condition on properties of the target. Simple conditions are considered true if the property of the current platform is the same as the associated value of the enum. For example, Condition::OS(OS::Windows) will be true on Windows and false otherwise. And, Or and Not variants provide logical operations on nested conditions. True and False variants provide conditions which are always true and false, respectively.

Variants
Arch(Arch)OS(OS)Family(Family)Env(Env)PointerWidth(PointerWidth)Endian(Endian)And(Vec<Condition>)Or(Vec<Condition>)Not(Box<Condition>)TrueFalse
Methods

impl Condition

fn eval(&self, target: &Target) -> bool

Evaluate the condition for target. Returns true if target matches the condition.

fn negate(&self) -> Condition

Construct a condition opposite to self.